# api/openapi.yaml
openapi: 3.1.0
info:
  title: ForgeERP API
  version: 0.1.0
  description: >
    ForgeERP modernized ERP/CRM API (Dolibarr functional successor).
    This document is authoritative for REST surface; phase routers must register
    every path declared here.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/reports/stock-valuation:
    get:
      operationId: stockValuation
      summary: PMP valuation for one warehouse (requires reporting.valuation.read)
      responses:
        '200': { description: Report }
        '400': { description: warehouse_id required }
...